generate `.env` with user supplied values

+++ src/setup/create_cron_script.py
@@ -9,10 +9,8 @@ DEST = "/etc/cron.weekly/system_email"
 def create_cron_script():
     content = read_cron_script_src()
 
-    if os.environ.get('dry_run'):
-        return
-
     write_file(content)
+
     set_permissions()
 
 
@@ -28,6 +26,9 @@ def read_cron_script_src():
 def write_file(content):
     print(f"create_cron_script.write_file - writing {DEST}")
 
+    if os.environ.get('opt_dry_run'):
+        return
+
     with open(DEST, 'w') as f:
         print(f"create_cron_script - writing script to {DEST}")
         f.write(content)
@@ -35,6 +36,9 @@ def write_file(content):
 
 def set_permissions():
     print(f"create_cron_script.set_permissions - Setting {DEST} permissions.")
+
+    if os.environ.get('opt_dry_run'):
+        return
 
     subprocess.run(
         ["chmod", "+x", DEST],
